Do not use AsyncTask because the core does things async anyway and move sending relayed things to new file (if it turns out that it's still too slow, then there are more useful APIs then AsyncApi).
Also, after sharing to multiple contacts, return to the standard view.
